backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "python ldap server"
authors = ["pavel_milanov <milanov71@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
ldap3 = "^2.9.1"
environs = "^9.5.0"
pycryptodome = "^3.17"
loguru = "^0.6.0"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
passlib = "^1.7.4"
redis = "^4.5.4"
tortoise-orm = "^0.19.3"
asyncpg = "^0.27.0"
aerich = "^0.7.1"
apscheduler = "^3.10.1"
aiofiles = "^23.1.0"
dramatiq = "^1.14.2"
websockets = "^11.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
ruff = "^0.0.288"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 90
exclude = [".venv", ".pytest_cache"]